# api/modules/linux/openapi.yaml
openapi: 3.1.0
...
info:
  title: Opale Linux fleet API
  version: 1.0.0
  description: |
    Control plane for Debian workstations managed by pull: devices enrol with a self-generated Ed25519
    key, are approved (or auto-approved by serial pre-registration), fetch their assignment at each
    check-in, pull their goal state from Opale's read-only git mirror with `ansible-pull`, and report
    the result. Opale never runs code on a Linux device: it only chooses which *signed* revision each
    ring applies.

    ## Device authentication (all `/api/linux/agent/*` operations except the git smart-HTTP routes)
    Every request carries:
    - `x-opale-key`: device key fingerprint = hex(sha256(raw 32-byte Ed25519 public key))
    - `x-opale-timestamp`: unix seconds
    - `x-opale-nonce`: 16 random bytes, hex
    - `x-opale-signature`: base64(Ed25519.sign(private_key, canonical))

    `canonical = "opale-linux-v1\n" + METHOD + "\n" + path_with_query + "\n" + timestamp + "\n" + nonce + "\n" + hex(sha256(raw_request_body))`
    (`path_with_query` is the request target as sent, e.g. `/api/linux/agent/checkin`; an empty body hashes as sha256 of zero bytes.)

    Server checks, in order: key row exists (on `/enroll` any status; elsewhere status must be `approved`),
    `|now - timestamp| <= 300 s` (else 401 `CLOCK_SKEW` with `server_time`), nonce unseen in the last 10
    minutes, signature valid over the exact bytes received. Failures are 401 with a `code`.

    ## Git smart HTTP
...

...
servers:
...
security:
  - deviceSignature: []
paths:
...
  /linux/devices:
    get:
      operationId: linuxListDevices
      tags: [linux-admin]
      summary: List pull-managed devices (server paged)
      security: [ { entraJwt: [] } ]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [approved, revoked] } }
        - { name: profile, in: query, schema: { type: string, maxLength: 64 } }
        - { name: ring, in: query, schema: { $ref: '#/components/schemas/Ring' } }
        - { name: lagging, in: query, schema: { type: boolean } }
        - { name: apply_status, in: query, schema: { $ref: '#/components/schemas/ApplyStatus' } }
        - { name: escrow, in: query, schema: { type: string, enum: [ok, missing] }, description: "`missing` = LUKS root detected but no current recovery key" }
        - { name: online, in: query, schema: { type: boolean }, description: "last_seen within 1 hour" }
        - { name: q, in: query, schema: { type: string, maxLength: 200 }, description: "hostname / serial / user substring" }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        '200':
          description: Page.
          content:
            application/json:
              schema:
                type: object
                required: [rows, total]
                properties:
                  rows: { type: array, items: { $ref: '#/components/schemas/LinuxDevice' } }
                  total: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /linux/devices/assign-bulk:
    post:
      operationId: linuxAssignBulk
      tags: [linux-admin]
      summary: Assign profile and/or ring to many devices
      security: [ { entraJwt: [] } ]
      description: Targets are `ids` or all pull-managed members of `group_id` (native group, recursive). Non-pull rows are skipped and counted. Audits `linux_assignment_changed` per device.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                ids: { type: array, items: { type: string, format: uuid }, maxItems: 500 }
                group_id: { type: string, format: uuid }
                profile: { $ref: '#/components/schemas/ProfileSlug' }
                ring: { $ref: '#/components/schemas/Ring' }
              anyOf:
                - { required: [ids] }
                - { required: [group_id] }
      responses:
        '200':
          description: Counts.
          content: { application/json: { schema: { $ref: '#/components/schemas/BulkResult' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /linux/devices/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      operationId: linuxGetDevice
      tags: [linux-admin]
      summary: Device detail
      security: [ { entraJwt: [] } ]
      responses:
        '200':
          description: Device.
          content: { application/json: { schema: { $ref: '#/components/schemas/LinuxDeviceDetail' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: linuxUpdateDevice
      tags: [linux-admin]
      summary: Change assignment (profile, ring, user)
      security: [ { entraJwt: [] } ]
      description: Takes effect at the device's next check-in. `assigned_user_id` goes through the shared inventory helper (audit `device_assigned`); profile/ring audit `linux_assignment_changed`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              minProperties: 1
              properties:
                profile: { $ref: '#/components/schemas/ProfileSlug' }
                ring: { $ref: '#/components/schemas/Ring' }
                assigned_user_id: { type: [string, 'null'], maxLength: 100, description: "users_cache.entra_id or null to unassign" }
      responses:
        '200':
          description: Updated device.
          content: { application/json: { schema: { $ref: '#/components/schemas/LinuxDeviceDetail' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /linux/devices/{id}/revoke:
    post:
      operationId: linuxRevokeDevice
      tags: [linux-admin]
      summary: Revoke the device's key (device row kept)
      security: [ { entraJwt: [] } ]
      x-opale-interactive: true
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                reason: { type: string, minLength: 5, maxLength: 500 }
      responses:
        '200':
          description: Revoked.
          content: { application/json: { schema: { $ref: '#/components/schemas/LinuxDeviceDetail' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: No approved key to revoke.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /linux/devices/{id}/reports:
    get:
      operationId: linuxListReports
      tags: [linux-admin]
      summary: Apply history of a device (newest first)
      security: [ { entraJwt: [] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ApplyStatus' } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        '200':
          description: Page.
          content:
            application/json:
              schema:
                type: object
                required: [rows, total]
                properties:
                  rows: { type: array, items: { $ref: '#/components/schemas/ApplyReport' } }
                  total: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /linux/devices/{id}/recovery-keys:
    get:
      operationId: linuxListRecoveryKeys
      tags: [linux-admin]
      summary: Recovery-key metadata (never the secret)
      security: [ { entraJwt: [] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Keys, current first.
          content:
            application/json:
              schema:
                type: object
                required: [rows]
                properties:
                  rows: { type: array, items: { $ref: '#/components/schemas/RecoveryKeyMeta' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /linux/devices/{id}/recovery-keys/{kid}/reveal:
    post:
      operationId: linuxRevealRecoveryKey
      tags: [linux-admin]
      summary: Reveal a recovery key (reason required, fail-closed audit)
      security: [ { entraJwt: [] } ]
      x-opale-interactive: true
      x-opale-rate-limit: "10/min"
      description: The audit row (`linux_recovery_key_viewed`, with the reason) and `last_viewed_*` are committed in one transaction BEFORE the plaintext is sent; if that commit fails the secret is not returned. Decryption happens in memory with the mounted escrow private key.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: kid, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                reason: { $ref: '#/components/schemas/Reason' }
      responses:
        '200':
          description: The secret. Display for at most 30 s client-side.
          content: { application/json: { schema: { $ref: '#/components/schemas/RevealedSecret' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500':
          description: DECRYPT_FAILED (key_id mismatch or unreadable escrow key; audited as failed) or AUDIT_FAILED (the audit + `last_viewed_*` transaction could not be committed; the secret is not returned).
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  /linux/dashboard:
    get:
      operationId: linuxDashboard
      tags: [linux-admin]
      summary: Fleet KPIs
      security: [ { entraJwt: [] } ]
      responses:
        '200':
          description: KPIs.
          content: { application/json: { schema: { $ref: '#/components/schemas/LinuxDashboard' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ─────────────────── INVENTORY ADDITIONS (registered by the inventory module) ───────────────────
  /devices/{id}:
    patch:
      operationId: deviceSetAssignedUser
      tags: [inventory]
      x-opale-module: inventory
      summary: Assign a user to any device (platform-neutral)
      security: [ { entraJwt: [] } ]
      description: The only manual writer of `devices.assigned_user_id`. Validated against `users_cache`. Audits `device_assigned` with before/after.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [assigned_user_id]
              properties:
                assigned_user_id: { type: [string, 'null'], maxLength: 100 }
      responses:
        '200':
          description: "{ id, assigned_user_id, assigned_user_name }"
          content:
            application/json:
              schema:
                type: object
                required: [id, assigned_user_id]
                properties:
                  id: { type: string, format: uuid }
                  assigned_user_id: { type: [string, 'null'] }
                  assigned_user_name: { type: [string, 'null'] }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    deviceSignature:
      type: apiKey
      in: header
      name: x-opale-signature
      description: Ed25519 signed request (see the top-level description); also requires x-opale-key, x-opale-timestamp, x-opale-nonce.
    gitToken:
      type: http
      scheme: bearer
      description: Short-lived git token from `/checkin` (`git.token`).
    entraJwt:
      type: http
      scheme: bearer
      description: Entra ID access token or `opl_` CLI token. Operations marked `x-opale-interactive` refuse CLI tokens (403 INTERACTIVE_ONLY).
  responses:
    BadRequest:
      description: Validation error.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Missing/invalid credential. Device codes — SIGNATURE_INVALID, UNKNOWN_KEY, CLOCK_SKEW (with server_time), NONCE_REPLAY, REVOKED, NOT_APPROVED; git — TOKEN_INVALID, TOKEN_EXPIRED.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Not admin, or INTERACTIVE_ONLY for CLI tokens.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Not found.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    TooManyRequests:
      description: Rate limited; `Retry-After` header and `retry_after_ms`.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: "French, user-facing" }
        code: { type: string, description: "Machine-readable, UPPER_SNAKE" }
        server_time: { type: integer, description: "unix seconds; present on CLOCK_SKEW" }
        retry_after_ms: { type: integer }
        details: { type: object, additionalProperties: true }
      examples:
        - { error: "Horloge du poste décalée", code: CLOCK_SKEW, server_time: 1790800000 }
        - { error: "Poste géré par état désiré", code: PULL_MANAGED }
...
    Fingerprint:
      type: string
      pattern: '^[0-9a-f]{64}$'
      description: hex(sha256(raw Ed25519 public key))
    Ring:
      type: string
      enum: [pilot, stable]
    ProfileSlug:
      type: string
      pattern: '^[a-z0-9][a-z0-9-]{0,63}$'
      description: "`profiles/<slug>.yml` in the fleet repo"
      examples: [ field-researcher, admin ]
...
    EnrollmentStatus:
      type: string
      enum: [pending, approved, rejected, revoked]
    ApplyStatus:
      type: string
      enum: [success, failed, partial, skipped]
    KeyBacking:
      type: string
      enum: [software, tpm]
    MirrorState:
      type: string
      enum: [absent, cloning, ready, stale, error, unavailable]
    Reason:
      type: object
      additionalProperties: false
      required: [category, note]
      properties:
        category: { type: string, enum: [maintenance, troubleshoot, audit, incident, other] }
        note: { type: string, minLength: 5, maxLength: 500 }
      examples:
        - { category: incident, note: "Utilisateur bloqué au démarrage, ticket #412" }

    # ── agent payloads ──
...
    UserRef:
      type: [object, 'null']
      required: [entra_id, display_name]
      properties:
        entra_id: { type: string }
        display_name: { type: [string, 'null'] }
        email: { type: [string, 'null'] }
    KeyInfo:
      type: [object, 'null']
      required: [id, fingerprint, key_backing, status]
      properties:
        id: { type: string, format: uuid }
        fingerprint: { $ref: '#/components/schemas/Fingerprint' }
        key_backing: { $ref: '#/components/schemas/KeyBacking' }
        status: { $ref: '#/components/schemas/EnrollmentStatus' }
        agent_version: { type: [string, 'null'] }
        os_version: { type: [string, 'null'] }
        last_seen_at: { type: [string, 'null'], format: date-time }
        approved_at: { type: [string, 'null'], format: date-time }
        approved_by: { type: [string, 'null'] }
        revoked_at: { type: [string, 'null'], format: date-time }
        revoked_by: { type: [string, 'null'] }
        revoke_reason: { type: [string, 'null'] }
    LinuxDevice:
      type: object
      required: [id, hostname, serial, platform, managed_by, profile, ring, assigned_user, os, last_seen, online, disk_used_pct, key, last_revision_applied, last_successful_revision, last_apply_status, last_apply_at, ring_tip, lagging, needs_escrow]
      properties:
        id: { type: string, format: uuid }
        hostname: { type: string }
        serial: { type: [string, 'null'] }
        platform: { type: string, const: linux }
        managed_by: { type: string, const: pull }
        profile: { type: [string, 'null'] }
        profile_in_repo: { type: [boolean, 'null'], description: "null when the mirror is not ready" }
        ring: { type: [string, 'null'] }
        assigned_user: { $ref: '#/components/schemas/UserRef' }
        os: { type: [string, 'null'] }
        last_seen: { type: [string, 'null'], format: date-time }
        online: { type: boolean }
        disk_used_pct: { type: [number, 'null'] }
        key: { $ref: '#/components/schemas/KeyInfo' }
        last_revision_applied: { type: [string, 'null'] }
        last_successful_revision: { type: [string, 'null'] }
        last_apply_status: { type: [string, 'null'] }
        last_apply_at: { type: [string, 'null'], format: date-time }
        ring_tip: { type: [string, 'null'] }
        lagging: { type: [boolean, 'null'] }
        needs_escrow: { type: boolean, description: "luks_root reported and no current recovery key under the current key_id" }
    LinuxDeviceDetail:
      allOf:
        - { $ref: '#/components/schemas/LinuxDevice' }
        - type: object
          required: [kernel, luks_root, last_report, laps, recovery_keys, converted_from_windows]
          properties:
            kernel: { type: [string, 'null'] }
            luks_root: { type: [boolean, 'null'] }
            last_report: { $ref: '#/components/schemas/ApplyReportOrNull' }
            laps:
              type: [object, 'null']
              description: same shape as GET /api/devices/:id `laps`
              required: [username, password_changed_at, rotation_requested_at, last_viewed_at, last_viewed_by_name]
              properties:
                username: { type: [string, 'null'] }
                password_changed_at: { type: [string, 'null'], format: date-time }
                rotation_requested_at: { type: [string, 'null'], format: date-time }
                last_viewed_at: { type: [string, 'null'], format: date-time }
                last_viewed_by_name: { type: [string, 'null'] }
            recovery_keys: { type: array, items: { $ref: '#/components/schemas/RecoveryKeyMeta' } }
            converted_from_windows: { type: boolean }
    ApplyReport:
      type: object
      required: [id, revision, status, started_at, finished_at, error_summary, log_tail, agent_version, received_at]
      properties:
        id: { type: string, format: uuid }
        revision: { type: [string, 'null'] }
        status: { $ref: '#/components/schemas/ApplyStatus' }
        started_at: { type: [string, 'null'], format: date-time }
        finished_at: { type: [string, 'null'], format: date-time }
        error_summary: { type: [string, 'null'] }
        log_tail: { type: [string, 'null'] }
        agent_version: { type: [string, 'null'] }
        received_at: { type: string, format: date-time }
    ApplyReportOrNull:
      anyOf:
        - { $ref: '#/components/schemas/ApplyReport' }
        - { type: 'null' }
    RecoveryKeyMeta:
      type: object
      required: [id, kind, label, key_id, created_at, superseded_at, current, last_viewed_at, last_viewed_by_name]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [luks_recovery, tpm_owner] }
        label: { type: string }
        key_id: { type: string }
        created_at: { type: string, format: date-time }
        superseded_at: { type: [string, 'null'], format: date-time }
        current: { type: boolean, description: "not superseded and key_id = current escrow key" }
        last_viewed_at: { type: [string, 'null'], format: date-time }
        last_viewed_by_name: { type: [string, 'null'] }
    RevealedSecret:
      type: object
      required: [kind, label, secret, key_id, created_at]
      properties:
        kind: { type: string, enum: [luks_recovery, tpm_owner] }
        label: { type: string }
        secret: { type: string }
        key_id: { type: string }
        created_at: { type: string, format: date-time }
...
    BulkResult:
      type: object
      required: [ok, skipped, errors]
      properties:
        ok: { type: integer }
        skipped: { type: integer }
        errors:
          type: array
          items:
            type: object
            required: [id, code]
            properties:
              id: { type: string, description: "row id, device id, or the row index for imports" }
              code: { type: string }
              message: { type: string }
      examples:
        - { ok: 18, skipped: 1, errors: [ { id: "3", code: HOSTNAME_TAKEN, message: "lx-dupont existe déjà" } ] }
...
    LinuxDashboard:
      type: object
      required: [devices_total, by_revision, lagging, offline, failed_applies, pending_approvals, not_escrowed, mirror_state]
      properties:
        devices_total: { type: integer }
        by_revision:
          type: array
          items:
            type: object
            required: [revision, ring, count]
            properties:
              revision: { type: [string, 'null'] }
              ring: { type: [string, 'null'] }
              count: { type: integer }
        lagging: { type: integer }
        offline:
          type: object
          required: [days, count]
          properties:
            days: { type: integer, description: "settings.agent_offline_days" }
            count: { type: integer }
        failed_applies: { type: integer, description: "devices whose last apply failed or is partial" }
        pending_approvals: { type: integer }
        not_escrowed: { type: integer }
        mirror_state: { $ref: '#/components/schemas/MirrorState' }